refactor(bienbot-ui): extract handleExecuteAction into useExecuteAction hook
Moves the ~213-line handleExecuteAction useCallback out of BienBotPanel.jsx
into a dedicated useExecuteAction hook. The hook owns executingActionId state
and executingActionRef, and exposes { handleExecuteAction, executingActionId }.
Removes the broadcastEvent import from BienBotPanel.jsx (now only used in the hook).
