+++ src/components/BienBotPanel/BienBotPanel.jsx
@@ -25,8 +25,9 @@ import {
 } from '../../utilities/bienbot-suggestions';
 import { logger } from '../../utilities/logger';
 import { decodeHtmlEntities } from '../../utilities/html-entities';
-import { eventBus, broadcastEvent } from '../../utilities/event-bus';
+import { eventBus } from '../../utilities/event-bus';
 import { useBienBotOptimistic } from './useBienBotOptimistic';
+import { useExecuteAction } from './useExecuteAction';
 import PendingActionsArea from './PendingActionsArea';
 import ChatInputArea from './ChatInputArea';
 import SessionHistoryView from './SessionHistoryView';
@@ -101,8 +102,6 @@ export default function BienBotPanel({
   const [attachment, setAttachment] = useState(null);
   const [attachmentPreviewUrl, setAttachmentPreviewUrl] = useState(null);
   const [shareOpen, setShareOpen] = useState(false);
-  const [executingActionId, setExecutingActionId] = useState(null);
-  const executingActionRef = useRef(null);
   // Prevents duplicate plan creation if the user double-clicks "Plan this" before
   // createPlan resolves (isStreaming/isLoading are only set later, inside sendHiddenMessage).
   const isPlanningRef = useRef(false);
@@ -182,6 +181,18 @@ export default function BienBotPanel({
     clearPersistedSession,
     fetchSessions
   } = useBienBot({ invokeContext, navigationSchema, userId: user?._id });
+
+  const { handleExecuteAction, executingActionId } = useExecuteAction({
+    pendingActions,
+    executeActions,
+    cancelAction,
+    navigate,
+    onClose,
+    broadcastOptimistic,
+    appendStructuredContent,
+    appendMessage,
+    userId: user?._id,
+  });
 
   // ── Focus trap: keep focus inside the dialog while open ────────────────
   useFocusTrap(panelRef, open);
@@ -784,221 +795,6 @@ export default function BienBotPanel({
   );
 
   // ── Action execute/cancel ─────────────────────────────────────────────────
-  const handleExecuteAction = useCallback(
-    async (actionId) => {
-      logger.debug('[BienBotPanel] Executing action', { actionId });
-
-      // Guard against rapid double-clicks: state updates are async so we
-      // use a ref to prevent re-entry before the disabled prop propagates.
-      if (executingActionRef.current === actionId) return;
-      executingActionRef.current = actionId;
-
-      // Check if it's a navigate action (client-only, no server call)
-      const action = pendingActions.find(a => (a._id || a.id) === actionId);
-      if (action && action.type === 'navigate_to_entity') {
-        const url = action.payload?.url;
-        // Validate URL contains real IDs, not LLM placeholders like <unknown> or <experienceId>
-        if (url && !/<[^>]+>/.test(url)) {
-          navigate(url);
-          cancelAction(actionId);
-          onClose();
-          executingActionRef.current = null;
-          return;
-        }
-        // Bad URL — cancel without navigating
-        cancelAction(actionId);
-        executingActionRef.current = null;
-        return;
-      }
-
-      // For select_plan, cancel all other select_plan actions (user picked one)
-      if (action && action.type === 'select_plan') {
-        const otherSelectPlans = pendingActions.filter(
-          a => a.type === 'select_plan' && (a._id || a.id) !== actionId
-        );
-        for (const other of otherSelectPlans) {
-          cancelAction(other._id || other.id);
-        }
-      }
-
-      // For select_destination, cancel all other select_destination actions (user picked one)
-      if (action && action.type === 'select_destination') {
-        const otherSelectDestinations = pendingActions.filter(
-          a => a.type === 'select_destination' && (a._id || a.id) !== actionId
-        );
-        for (const other of otherSelectDestinations) {
-          cancelAction(other._id || other.id);
-        }
-      }
-
-      // Show executing state on the action card
-      setExecutingActionId(actionId);
-
-      // ─── Optimistic UI updates ──────────────────────────────────────────────
-      // Fire BEFORE the API call so DOM elements already on screen reflect the
-      // change instantly — same pattern as the manual UI interactions.
-      // All events are flagged _optimistic:true. There is no rollback: if the
-      // server call fails the stale optimistic state persists until the next
-      // data refresh (acceptable for the BienBot confirmed-action flow).
-      //
-      // Strategy by action category:
-      //  • Plan item changes  → plan:operation (UPDATE_ITEM / DELETE_ITEM via
-      //    usePlanManagement's CRDT handler) + plan:item:updated for the open modal
-      //  • Plan-level deletes → plan:deleted (direct state removal)
-      //  • Plan-level updates → plan:operation UPDATE_PLAN / REORDER_ITEMS
-      //  • Completion toggles → plan:item:completed / plan:item:uncompleted
-      //    (dedicated direct-patch handler in usePlanManagement)
-      //  • Social follows     → follow:created / follow:deleted / follow:request:accepted
-      //  • Actions that need the full server response to construct a meaningful
-      //    optimistic state (create_*, sync_plan, costs, notes/details) are skipped.
-      broadcastOptimistic(action, user?._id);
-
-      const result = await executeActions([actionId]);
-
-      setExecutingActionId(null);
-      // On success the action is removed from pendingActions and the card
-      // disappears on the next render — leave the ref set so any click that
-      // snuck in while the fetch was in-flight (already queued as a macrotask)
-      // is still blocked. On failure the card stays visible and the user must
-      // be able to retry, so we clear the ref only then.
-      if (!result) {
-        executingActionRef.current = null;
-      }
-
-      // Build a feedback message summarizing what happened
-      if (result?.results) {
-        const feedbackLines = [];
-        for (const actionResult of result.results) {
-          if (actionResult.success) {
-            if (actionResult.type === 'create_plan') {
-              const expName = actionResult.result?.experience?.name || '';
-              const itemCount = actionResult.result?.plan?.length || 0;
-              feedbackLines.push(
-                `\u2705 Plan created${expName ? ` for ${expName}` : ''}${itemCount > 0 ? ` with ${itemCount} item${itemCount !== 1 ? 's' : ''}` : ''}. Taking you there now\u2026`
-              );
-            } else if (actionResult.type === 'update_plan_item') {
-              // For plan item updates, summarize what changed
-              const payload = action?.payload || {};
-              const changes = [];
-              if (payload.scheduled_date) changes.push('scheduled date');
-              if (payload.scheduled_time) changes.push('scheduled time');
-              if (payload.text) changes.push('description');
-              if (payload.cost !== undefined) changes.push('cost');
-              if (payload.activity_type) changes.push('activity type');
-              if (payload.complete !== undefined) changes.push(payload.complete ? 'marked complete' : 'marked incomplete');
-              if (payload.location) changes.push('location');
-              const summary = changes.length > 0 ? changes.join(', ') : 'details';
-              feedbackLines.push(`\u2705 Plan item updated: ${summary}`);
-            } else if (actionResult.type === 'mark_plan_item_complete' || actionResult.type === 'mark_plan_item_incomplete') {
-              const isComplete = actionResult.type === 'mark_plan_item_complete';
-              const itemPayload = action?.payload || {};
-              const itemIdStr = itemPayload.item_id?.toString ? itemPayload.item_id.toString() : itemPayload.item_id;
-              const planItems = Array.isArray(actionResult.result?.plan) ? actionResult.result.plan : [];
-              const matchedItem = planItems.find(i => (i._id?.toString ? i._id.toString() : i._id) === itemIdStr);
-              const itemName = matchedItem?.text || action?.description || '';
-              feedbackLines.push(isComplete
-                ? `\u2705 ${itemName ? `"${itemName}" marked complete` : 'Plan item marked complete'}`
-                : `\u2705 ${itemName ? `"${itemName}" marked incomplete` : 'Plan item marked incomplete'}`
-              );
-            } else {
-              const entityName = actionResult.result?.name || actionResult.result?.title || actionResult.result?.content || '';
-              const rawLabel = (actionResult.type || '').replace(/_/g, ' ');
-              const typeLabel = rawLabel.charAt(0).toUpperCase() + rawLabel.slice(1);
-              feedbackLines.push(`\u2705 ${typeLabel}${entityName ? `: ${entityName}` : ''}`);
-            }
-          } else {
-            const rawLabel = (actionResult.type || '').replace(/_/g, ' ');
-            const typeLabel = rawLabel.charAt(0).toUpperCase() + rawLabel.slice(1);
-            feedbackLines.push(`\u274c ${typeLabel}: ${actionResult.error || 'failed'}`);
-          }
-        }
-        if (feedbackLines.length > 0) {
-          appendMessage({
-            _id: `exec-result-${Date.now()}`,
-            role: 'assistant',
-            content: feedbackLines.join('\n'),
-            createdAt: new Date().toISOString(),
-            isActionResult: true
-          });
-        }
-
-        // Auto-navigate to newly created entities (panel stays open)
-        let navigated = false;
-        for (const actionResult of result.results) {
-          const navUrl = getNavigationUrlForResult(actionResult);
-          if (navUrl) {
-            navigate(navUrl);
-            navigated = true;
-            break; // Only navigate once
-          }
-        }
-
-        // Re-emit entity events after navigation so that page components
-        // that mount on the new route can pick them up (they may have missed
-        // the initial broadcast fired before navigate() was called).
-        if (navigated && result.results) {
-          setTimeout(() => {
-            try {
-              for (const actionResult of result.results) {
-                if (!actionResult.success) continue;
-                const entity = actionResult.result || actionResult.entity || actionResult.data;
-                if (!entity) continue;
-                switch (actionResult.type) {
-                  case 'create_plan':
-                    broadcastEvent('plan:created', {
-                      plan: entity,
-                      planId: entity._id,
-                      experienceId: entity.experience?._id || entity.experience,
-                      version: Date.now()
-                    });
-                    break;
-                  case 'update_plan':
-                  case 'add_plan_items':
-                  case 'sync_plan':
-                    broadcastEvent('plan:updated', {
-                      plan: entity,
-                      planId: entity._id || actionResult.planId,
-                      version: Date.now()
-                    });
-                    break;
-                  case 'create_experience':
-                    broadcastEvent('experience:created', { experience: entity, experienceId: entity._id });
-                    break;
-                  case 'update_experience':
-                    broadcastEvent('experience:updated', { experience: entity, experienceId: entity._id });
-                    break;
-                  case 'create_destination':
-                    broadcastEvent('destination:created', { destination: entity, destinationId: entity._id });
-                    break;
-                  case 'update_destination':
-                    broadcastEvent('destination:updated', { destination: entity, destinationId: entity._id });
-                    break;
-                  default:
-                    break;
-                }
-              }
-            } catch (e) { /* silently ignore */ }
-          }, 0);
-        }
-      }
-
-      // Contextual enrichment: suggestions/tips/photos after entity creation
-      if (result?.enrichment) {
-        appendStructuredContent(result.enrichment);
-      }
-
-      // Post-execution follow-up: LLM "what's next?" message with plan items context
-      if (result?.followUpMessage) {
-        appendMessage({
-          _id: `exec-followup-${Date.now()}`,
-          role: 'assistant',
-          content: result.followUpMessage,
-          createdAt: new Date().toISOString()
-        });
-      }
-    },
-    [executeActions, pendingActions, cancelAction, navigate, appendStructuredContent, appendMessage, onClose]
-  );
 
   const handleCancelAction = useCallback(
     (actionId) => {
